pyproject.toml
[tool.poetry]
name = "flask-mongo-jwt"
version = "0.1.0"
description = ""
authors = ["Robert Alexander <gogonegro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
Flask = "^2.3.3"
Flask-Cors = "^4.0.0"
itsdangerous = "^2.1.2"
Jinja2 = "^3.1.2"
MarkupSafe = "^2.1.3"
pymongo = "^4.5.0"
six = "^1.16.0"
werkzeug = "^2.3.7"
flask-bcrypt = "^1.0.1"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
